Reorganizes files into src and __test
