+++ __test/lib/values.test.js
@@ -3,7 +3,7 @@ const {
   isCustomProperty,
   roundToPrecision,
   splitCssUnit,
-} = require('../values');
+} = require('../../src/lib/values');
 
 /**
  * Compare numerical strings.
