+++ __test/sharedConfigs.js
