+++ src/lib/cleanClone.js
