+++ src/lib/detectCalcWrapper.js
