+++ src/tidy-function.js
