+++ src/tidy-var.js
