--- test/sourcemap/function-span.css
@@ -1,3 +1,0 @@
-div {
-	width: tidy-span(2);
-}
