--- test/sourcemap/span.css
@@ -1,3 +1,0 @@
-div {
-	tidy-span: 2;
-}
