--- test/sourcemap/span.generated.css
@@ -1,4 +1,0 @@
-div {
-	width: calc((((100vw - 0.625rem * 2) / 12 - 1.1458rem) * 2) + 1.25rem);
-	max-width: calc((((90rem - 0.625rem * 2) / 12 - 1.1458rem) * 2) + 1.25rem);
-}
